requirements.txt
appdirs==1.4.4
asgiref==3.2.7
blis==0.4.1
catalogue==1.0.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9.2
cycler==0.10.0
cymem==2.0.3
DateTime==4.3
decorator==4.4.2
distlib==0.3.1
Django==3.0.5
dnspython==1.16.0
doc2text==0.2.4
docx==0.2.4
en-core-web-sm==2.2.5
filelock==3.0.12
future==0.18.2
idna==2.9
image==1.5.28
importlib-metadata==1.6.0
joblib==0.14.1
kiwisolver==1.2.0
lxml==4.5.0
matplotlib==3.2.1
mime==0.1.0
murmurhash==1.0.2
mysql-connector-python==8.0.20
nltk==3.5
numpy==1.18.2
pandas==1.0.3
Pillow==7.1.1
plac==1.1.3
preshed==3.0.2
pycparser==2.20
pygame==1.9.6
pymongo==3.10.1
pyparsing==2.4.7
PyPDF2==1.26.0
pytesseract==0.3.4
python-dateutil==2.8.1
python-docx==0.8.10
python-telegram-bot==12.8
pytz==2019.3
PyYAML==5.3.1
regex==2020.4.4
regexp==0.1
requests==2.23.0
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
spacy==2.2.4
sqlparse==0.3.1
srsly==1.0.2
thinc==7.4.0
tornado==6.0.4
tqdm==4.45.0
urllib3==1.25.9
virtualenv==20.0.25
wasabi==0.6.0
xlrd==1.2.0
zipp==3.1.0
zope.interface==5.1.0
